Refactor ResourceProvider components.
  - Replace width/height parameter pairs with math::Size.
  - Introduce ImageDataDescriptor to package all image data settings.
  - Move ResourceProvider subclasses outside of ResourceProvider.

This is in anticipation of introducing new ResourceProvider methods
required by video.

+++ cobalt/math/size.h
@@ -22,7 +22,9 @@ class Size : public SizeBase<Size, int> {
 
   ~Size() {}
 
-  operator SizeF() const { return SizeF(width(), height()); }
+  operator SizeF() const {
+    return SizeF(static_cast<float>(width()), static_cast<float>(height()));
+  }
 
   std::string ToString() const;
 };
